Some further reorganising
