+++ packages/apps/tools/src/components/Global/ModuleExplorer/types.ts
@@ -1,4 +1,5 @@
 import type { Network } from '@/constants/kadena';
+import type { IncompleteModuleModel } from '@/hooks/use-module-query';
 import type { ChainwebChainId } from '@kadena/chainweb-node-client';
 import type {
   ContractCapability,
@@ -19,3 +20,14 @@ export type Outline =
   | ContractInterface
   | ContractCapability
   | ContractFunction;
+
+export const isModule = (
+  item: IncompleteModuleModel | Outline,
+): item is IncompleteModuleModel => {
+  return (
+    typeof item !== 'string' &&
+    (item as IncompleteModuleModel).chainId !== undefined &&
+    (item as IncompleteModuleModel).name !== undefined &&
+    (item as IncompleteModuleModel).networkId !== undefined
+  );
+};
